Handle issuetypes cache in new JiraSystemConfigLoader
