+++ main.py
@@ -16,9 +16,9 @@ if __name__ == '__main__':
     elif jira_options.action == 'me-as-assignee':
         print(jira.get_current_user_as_assignee())
     elif jira_options.action == 'fetch-issuetypes':
-        jira.update_issuetypes_cache()
+        jira.system_config_loader.update_issuetypes_cache()
         print('Updated local cache for issuetypes:')
-        pprint(jira.get_issuetypes_names_from_cache())
+        pprint(jira.system_config_loader.get_issuetypes_names_from_cache())
     elif jira_options.action == 'get-issue':
         for issue_key in jira_options.issues:
             issue = jira.issues.get(key=issue_key)
