+++ mantis/jira/__init__.py
@@ -2,4 +2,5 @@ from .jira_options import JiraOptions, parse_args
 from .jira_auth import JiraAuth
 from .jira_client import JiraClient
 from .jira_issues import JiraIssues
-from .utils import fetch_enums
+from .utils import JiraSystemConfigLoader
+
