+++ mantis/jira/utils/__init__.py
@@ -1,1 +1,2 @@
-from .fetch_enums import fetch_enums
+from .jira_system_config_loader import JiraSystemConfigLoader
+
